api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sentence-transformers==3.3.1
chromadb==0.5.23
groq==0.12.0
python-dotenv==1.0.1
pydantic==2.10.3
